clean up tags sdk lib
