+++ builtin/providers/aws/tags.go
@@ -3,8 +3,8 @@ package aws
 import (
 	"log"
 
-	"github.com/hashicorp/aws-sdk-go/aws"
-	"github.com/hashicorp/aws-sdk-go/gen/ec2"
+	"github.com/awslabs/aws-sdk-go/aws"
+	"github.com/awslabs/aws-sdk-go/service/ec2"
 	"github.com/hashicorp/terraform/helper/schema"
 )
 
@@ -19,18 +19,18 @@ func tagsSchema() *schema.Schema {
 
 // setTags is a helper to set the tags for a resource. It expects the
 // tags field to be named "tags"
-func setTags(conn *ec2.EC2, d *schema.ResourceData) error {
+func setTagsSDK(conn *ec2.EC2, d *schema.ResourceData) error {
 	if d.HasChange("tags") {
 		oraw, nraw := d.GetChange("tags")
 		o := oraw.(map[string]interface{})
 		n := nraw.(map[string]interface{})
-		create, remove := diffTags(tagsFromMap(o), tagsFromMap(n))
+		create, remove := diffTagsSDK(tagsFromMapSDK(o), tagsFromMapSDK(n))
 
 		// Set tags
 		if len(remove) > 0 {
 			log.Printf("[DEBUG] Removing tags: %#v", remove)
-			err := conn.DeleteTags(&ec2.DeleteTagsRequest{
-				Resources: []string{d.Id()},
+			_, err := conn.DeleteTags(&ec2.DeleteTagsInput{
+				Resources: []*string{aws.String(d.Id())},
 				Tags:      remove,
 			})
 			if err != nil {
@@ -39,8 +39,8 @@ func setTags(conn *ec2.EC2, d *schema.ResourceData) error {
 		}
 		if len(create) > 0 {
 			log.Printf("[DEBUG] Creating tags: %#v", create)
-			err := conn.CreateTags(&ec2.CreateTagsRequest{
-				Resources: []string{d.Id()},
+			_, err := conn.CreateTags(&ec2.CreateTagsInput{
+				Resources: []*string{aws.String(d.Id())},
 				Tags:      create,
 			})
 			if err != nil {
@@ -55,7 +55,7 @@ func setTags(conn *ec2.EC2, d *schema.ResourceData) error {
 // diffTags takes our tags locally and the ones remotely and returns
 // the set of tags that must be created, and the set of tags that must
 // be destroyed.
-func diffTags(oldTags, newTags []ec2.Tag) ([]ec2.Tag, []ec2.Tag) {
+func diffTagsSDK(oldTags, newTags []*ec2.Tag) ([]*ec2.Tag, []*ec2.Tag) {
 	// First, we're creating everything we have
 	create := make(map[string]interface{})
 	for _, t := range newTags {
@@ -63,7 +63,7 @@ func diffTags(oldTags, newTags []ec2.Tag) ([]ec2.Tag, []ec2.Tag) {
 	}
 
 	// Build the list of what to remove
-	var remove []ec2.Tag
+	var remove []*ec2.Tag
 	for _, t := range oldTags {
 		old, ok := create[*t.Key]
 		if !ok || old != *t.Value {
@@ -72,14 +72,14 @@ func diffTags(oldTags, newTags []ec2.Tag) ([]ec2.Tag, []ec2.Tag) {
 		}
 	}
 
-	return tagsFromMap(create), remove
+	return tagsFromMapSDK(create), remove
 }
 
 // tagsFromMap returns the tags for the given map of data.
-func tagsFromMap(m map[string]interface{}) []ec2.Tag {
-	result := make([]ec2.Tag, 0, len(m))
+func tagsFromMapSDK(m map[string]interface{}) []*ec2.Tag {
+	result := make([]*ec2.Tag, 0, len(m))
 	for k, v := range m {
-		result = append(result, ec2.Tag{
+		result = append(result, &ec2.Tag{
 			Key:   aws.String(k),
 			Value: aws.String(v.(string)),
 		})
@@ -89,7 +89,7 @@ func tagsFromMap(m map[string]interface{}) []ec2.Tag {
 }
 
 // tagsToMap turns the list of tags into a map.
-func tagsToMap(ts []ec2.Tag) map[string]string {
+func tagsToMapSDK(ts []*ec2.Tag) map[string]string {
 	result := make(map[string]string)
 	for _, t := range ts {
 		result[*t.Key] = *t.Value
